Fix quantile summary bugs
Note that the underlying algorithm for quantiles is not mergeable,
hence, it requires a linear strategy. However, precisely because it is
not generally mergeable, and it is also expensive, Prometheus generally
recommends using histograms on the client side, and it is left to the
Prometheus server to compute, merge, and aggregate, histograms,
recreating any required quantiles.

It is left for the future to implement mergeable summaries, an idea
would be a DDSketch algorithm or any variant thereof.